bug fixed

+++ src/components/three-scenes/star-canvas.tsx
@@ -3,13 +3,13 @@
 import React, { useRef, Suspense } from "react";
 import { Canvas, useFrame } from "@react-three/fiber";
 import { Points, PointMaterial, Preload } from "@react-three/drei";
-import * as random from "maath/random/dist/maath-random.esm";
+import { random } from "maath";
 
 interface StarsProps {}
 
 const Stars: React.FC<StarsProps> = () => {
   const ref = useRef<THREE.Points>(null);
-  const sphere = random.inSphere(new Float32Array(5000), { radius: 1.2 });
+  const sphere = random.inSphere(new Float32Array(5000), { radius: 1.2 }) as Float32Array;
 
   useFrame((state, delta) => {
     if (ref.current) {
